# Conflicts:
#	real_time/surface_src/java_src/Alice/src/main/java/com/roboclub/robobuggy/main/RobobuggyConfigFile.java
#	real_time/surface_src/java_src/Alice/src/main/java/com/roboclub/robobuggy/main/Robot.java
#	real_time/surface_src/java_src/Alice/src/main/java/com/roboclub/robobuggy/ui/GpsPanel.java
#	real_time/surface_src/java_src/Alice/src/main/java/com/roboclub/robobuggy/ui/Gui.java

+++ real_time/surface_src/java_src/Alice/src/main/java/com/roboclub/robobuggy/ui/AngleGraph.java
@@ -1,12 +1,11 @@
 package com.roboclub.robobuggy.ui;
 
-import java.awt.Color;
-import java.awt.Dimension;
-import java.awt.Graphics;
-
 import javax.swing.JLabel;
 import javax.swing.JPanel;
 import javax.swing.JTextField;
+import java.awt.Color;
+import java.awt.Dimension;
+import java.awt.Graphics;
 
 /**
  * {@link RobobuggyGUIContainer} used to display an angular graph
@@ -36,9 +35,9 @@ public class AngleGraph extends RobobuggyGUIContainer {
 	 * Update the {@link AngleGraph} with a new value
 	 * @param angle new value to display
 	 */
-	public void updateGraph(int angle) {
+	public void updateGraph(double angle) {
 		this.graph.updateGraph(angle);
-		this.reading.setText(Integer.toString(angle));
+		this.reading.setText(Double.toString(angle));
 	}
 	
 	/**
@@ -91,7 +90,7 @@ public class AngleGraph extends RobobuggyGUIContainer {
 			g.drawLine(startX, startY, endx, endy);
 		}
 		
-		public void updateGraph(int angle) {
+		public void updateGraph(double angle) {
 			endx = startX - (int)(radius * Math.sin(Math.toRadians(angle)));
 			endy = startY - (int)(radius * Math.cos(Math.toRadians(angle)));
 			Gui.getInstance().fixPaint();
